requirements.txt
numpy==1.23.5
opencv-python==4.8.0.76
pycocotools
matplotlib
absl-py
apache_beam
einops==0.6.1
Polygon3==3.0.9.1
pyclipper==1.3.0.post5
python-Levenshtein
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.2
timm
tqdm